requirements.txt
# native dependencies
Pillow==8.1.0
lxml==4.6.2
defusedxml==0.6.0
psycopg2==2.8.6
Django==2.2.16

# Other
amqp==5.0.5
pyyaml>=4.2b1
beautifulsoup4==4.8.0
httplib2<0.19.1
hyperlink==21.0.0
idna>=2.5,<2.11
urllib3==1.26.3
Paver==1.3.4
python-slugify==4.0.1
decorator==4.4.2
celery==5.0.5
kombu==5.0.2
vine==5.0.0
boto3==1.17.12
tqdm==4.57.0
Deprecated==1.2.11
wrapt==1.12.1
jsonfield==3.1.0
jsonschema==3.2.0
pyrsistent==0.17.3

# Django Apps
django-allauth==0.44.0
django-appconf==1.0.4
django-filter==2.4.0
django-imagekit==4.0.2
django-jsonfield==1.4.1
django-jsonfield-compat==0.4.4
django-taggit==1.3.0
django-markdownify==0.8.2
django-mptt==0.11.0
django-modeltranslation>=0.11,<0.17.0
django-treebeard==4.5
django-guardian<2.3.1
django-downloadview==2.1.1
django-polymorphic==3.0.0
django-tastypie<0.15.0
django-tinymce==3.2.0
django-grappelli==2.14.3
django-uuid-upload-path==1.0.0
oauthlib==3.1.0
pyopenssl==20.0.1

# geopython dependencies
pyproj==3.0.0.post1
OWSLib==0.23.0
pycsw==2.6.0
SQLAlchemy==1.3.23 # required by PyCSW
Shapely==1.7.1
mercantile==1.1.6
geoip2==4.1.0
numpy==1.19.*

# # Apps with packages provided in GeoNode's PPA on Launchpad.

# Django Apps
dj-database-url==0.5.0
dj-pagination==2.5.0
django-floppyforms<1.10.0
django-forms-bootstrap<=3.1.0
django-autocomplete-light==3.5.1
django-leaflet==0.27.1
django-invitations<1.9.4
django-recaptcha==2.0.6

# REST
djangorestframework==3.11.2
djangorestframework-gis==0.17
djangorestframework-guardian==0.3.0
drf-extensions==0.7.0
drf-writable-nested==0.6.2
drf-spectacular==0.13.2
dynamic-rest==2.0.0
Markdown==3.3.3

pinax-notifications==6.0.0
pinax-ratings==4.0.0

# GeoNode org maintained apps.
django-geonode-mapstore-client>=2.1.0
#-e git+https://github.com/GeoNode/geonode-mapstore-client.git@master#egg=django_geonode_mapstore_client
geonode-avatar==5.0.7
django-geonode-client==1.0.9
geonode-oauth-toolkit==2.1.1
geonode-user-messages==2.0.2
geonode-announcements==2.0.2
geonode-django-activity-stream==0.9.0
gn-arcrest==10.5.4
geonode-dialogos==3.0.1
geoserver-restconfig==2.0.4.8
gn-gsimporter==2.0.1
gisdata==0.5.4

# haystack/elasticsearch
django-haystack==3.0
elasticsearch>=2.0.0,<8.0.0

# datetimepicker widget
django-bootstrap3-datetimepicker-2==2.8.2

# AWS S3 dependencies
django-storages==1.11.1

# Django Caches
python-memcached<=1.59

# Contribs
xmltodict<0.12.1
inflection==0.3.1
jdcal==1.4.1
mock<5.0.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
timeout-decorator==0.5.0
python_resize_image==1.1.19

# required by monitoring
psutil==5.8.0
django-cors-headers==3.7.0
user-agents
xmljson
django-ipware<3.1
# no version here, use latest one with fresh data
pycountry

# production
uWSGI==2.0.19.1
gunicorn==20.0.4
ipython==7.20.0
docker==4.4.3
invoke==1.5.0

# tests
coverage==5.4
parse-type==0.5.2
requests-toolbelt==0.9.1
flake8==3.8.4
pytest==6.2.2
pytest-bdd==4.0.2
splinter==0.14.0
pytest-splinter==3.3.1
pytest-django==4.1.0
setuptools==53.0.0
pip==21.2.4
Twisted==20.3.0
factory-boy==3.2.0
flaky==3.7.0

# wfapp
django-translation-flags==1.0.6
django-ckeditor==6.0.0
django_extensions==3.1.0
django_libs==2.0.3
simplejson==3.17.2
pytexit==0.3.4
#CMS
wagtail==2.11.3 
wagtailtrans==2.1

# for reports
pandas==1.3.1
highchartexport==0.0.5
fpdf2==2.4.2
matplotlib
scipy